Tibble of converters and parsers to download data

Tibble of converters and parsers to download data. To update watchdog your may proceed in the following steps:

watchdog

Format

A data frame with 7 columns

url

url to download and parse, may be empty

file_raw

name of the raw file after download, may be empty

file_main

name of the main file after parsing, usually csv

processing

name of the processing function

univariate

TRUE for univariate and FALSE for multivariate

frequency

observations per year

comment

comment in free form

Details

1. Set the working directory to the package folder.

2. Update the watchdog dataset in memory.

3. usethis::use_data(watchdog)

4. devtools::check()

5. upload package on github